test: add unit tests for formatDate utility
- extracted date functions to src/lib/date-utils.ts to isolate them from broken dependencies
- added src/lib/date-utils.test.ts using node:test runner
- updated src/lib/utils.ts to re-export date functions
